ignore ssl socket errors that are caused by the ssl buffer
it is possible that there has been incoming data on the file descriptor,
but that is not ready to be recv'd by the ssl socket because the ssl
frame was incomplete. In that case we get an SSL_ERROR_WANT_READ error,
which we ignore and we go back to the pending/select call.